chore: factor out blob and file into separate libs
