--- packages/ipfs-core-utils/src/files/file.browser.js
@@ -1,4 +1,0 @@
-'use strict'
-/* eslint-env browser */
-
-exports.File = File
